+++ src/app/Models/Tag.php
@@ -4,12 +4,15 @@ namespace Backpack\NewsCRUD\app\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Backpack\CRUD\CrudTrait;
+use Cviebrock\EloquentSluggable\SluggableInterface;
+use Cviebrock\EloquentSluggable\SluggableTrait;
 
-class Tag extends Model
+class Tag extends Model implements SluggableInterface
 {
     use CrudTrait;
+    use SluggableTrait;
 
-     /*
+    /*
     |--------------------------------------------------------------------------
     | GLOBAL VARIABLES
     |--------------------------------------------------------------------------
@@ -22,6 +25,12 @@ class Tag extends Model
     protected $fillable = ['name'];
     // protected $hidden = [];
     // protected $dates = [];
+    protected $sluggable = [
+        'build_from' => 'slug_or_name',
+        'save_to'    => 'slug',
+        'on_update'  => true,
+        'unique'     => true,
+    ];
 
     /*
     |--------------------------------------------------------------------------
@@ -51,6 +60,16 @@ class Tag extends Model
     | ACCESORS
     |--------------------------------------------------------------------------
     */
+
+    // The slug is created automatically from the "name" field if no slug exists.
+    public function getSlugOrNameAttribute()
+    {
+        if ($this->slug != '') {
+            return $this->slug;
+        }
+
+        return $this->name;
+    }
 
     /*
     |--------------------------------------------------------------------------