+++ CHANGELOG.md
@@ -22,6 +22,15 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Nothing
 
 
+## 2.0.0 - 2016-07-24
+
+### Added
+- slugs for categories and tags;
+- check column for articles featured column;
+- switched to the 'admin' middleware for the routes;
+- removed setFeaturedAttribute on Article, because it was too specific;
+
+
 ## 1.0.3 - 2016-07-24
 
 ### Added